Latest/LastSyncRevision now are all long, refactored DeletedNotes
DeletedNotes of ISyncServer should only carry a list of Guids, not the
whole notes (as this would force clients to keep saved copies of already
deleted notes)

+++ Tomboy-library/Tomboy/Sync/Filesystem/FilesystemSyncClient.cs
@@ -64,7 +64,7 @@ namespace Tomboy.Sync.Filesystem
 		public Engine Engine {
 			get; private set;
 		}
-		public int LastSynchronizedRevision {
+		public long LastSynchronizedRevision {
 			get {
 				return manifest.LastSyncRevision;
 			}
